pyproject.toml
[tool.poetry]
name = "clonedncopied-bot"
version = "0.1.0"
description = ""
authors = ["Jonathan Miller <jon@jmill.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.1"
python-dotenv = "^0.17.0"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
ipykernel = "^5.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = 'clonedncopied_bot.bot:start'